moderation pages backend connection
Tag selection on moderation-not-jury page is connected with backend with 2 API requests (moderators/my-tags) GET and POST. GET request is executed at every render of the page and POST request is executed every change on tag list. If a user selects more than 5 tags or selects no tags, no request is sent and warning message is shown. The information of member being moderator or not is also taken from backend using profile/myProfile endpoint and necessary page is shown accordingly. On the moderation-not-moderator page when a member clicks apply button a POST request is sent to /moderators/request-promotion end point and a success message to user is shown. Necessary CSS updates of new components are also done.