backend connection updates
I was getting error message from the /moderators/my-tags POST endpoint because I was trying to send all of the tag information,however only the last changed one was expected. It is fixed in this commit. Also now I am rendering the posts with another GET request to endpoint /moderators/my-requests after a change occurs on the tags.