Unit tests for not-moderator page
I have written unit tests for rendering the non-moderator view on moderation page and applying for becoming a moderator with a button click. It simulates a mock API call to see if button works. I have also added some undefined respnse checks to moderator page and I needed add a dependency to package.json files as a warning message suggested.